home_18/app/requirements.txt
aniso8601==9.0.1
attrs==24.2.0
blinker==1.8.2
cffi==1.17.1
click==8.1.7
cryptography==43.0.1
Flask==3.0.3
flask-restx==1.3.0
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
importlib_resources==6.4.4
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.5
marshmallow==3.22.0
packaging==24.1
pycparser==2.22
PyJWT==2.9.0
pytz==2024.1
referencing==0.35.1
rpds-py==0.20.0
SQLAlchemy==2.0.34
typing_extensions==4.12.2
Werkzeug==3.0.4
